fix migration

+++ database/migrations/2025_06_25_032319_create_jurnal_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('jurnal', function (Blueprint $table) {
             $table->id('kd_jurnal');
-            $table->integer('kd_karyawan');
+            $table->unsignedBigInteger('kd_karyawan');
             $table->string('tanggal', 20);
             $table->string('jam', 20);
             $table->text('isi_jurnal');
